Refactor rendering functions
